SMASH

Documentation for SMASH.

SMASH.fractionalToCartesian

SMASH.minimd

SMASH.minimumVolumeEllipsoid

SMASH.fractionalToCartesian — Method

fractionToCartesian(a,unitcell)

Fractional to Cartesian coordinates for vector a.

SMASH.minimd — Method

minimimum distance image convention; using a unitcell=3x3, with a and b vectors in fractional coords

SMASH.minimumVolumeEllipsoid — Method

minimumVolumeEllipsoid(points; tolerance=1e-3, verbose::Bool=true)

Calculates the minimum volume ellipsoid for the submitted point cloud. Dimension of problem and Number of points discovered by inspection of point cloud.

This method uses the Khachiyan algorithm, solving the dual problem in N+1 dimensions.

Khachiyan 1996 is pretty impenetrable: http://dx.doi.org/10.1006/jagm.1996.0062 But these slides (see page 17) are fairly understandable: https://people.orie.cornell.edu/miketodd/ublndeta.pdf

here I use variable names as in the above talk.

It was implemented with considerable assistance by reading J.Cumby's source code for PIEFACE: https://github.com/jcumby/PIEFACE/blob/372b6ff6166e4996d86084a3116a8b606c25acfa/pieface/ellipsoid.py#L54-L71 Nb: Python/Numpy has crazy definiton of dotproduct = matrix multiplication. 8-[

Cumby's Python implementation seems heavily influenced by Nina Moshtagh's Matlab implementation: https://uk.mathworks.com/matlabcentral/fileexchange/9542-minimum-volume-enclosing-ellipsoid https://doi.org/10.1.1.116.7691

My implementation follows Moshtagh closely, as I find the code well documented & the Matlab syntax is very close to Julia.